board관련 클래스 분리 및 문서화
