--- src/main/java/com/jjbin/mysite/api/domain/Category.java
@@ -1,5 +1,0 @@
-package com.jjbin.mysite.api.domain;
-
-public enum Category {
-    A, B, C
-}
