+++ src/main/java/com/jjbin/mysite/api/repository/board/LikeRepository.java
@@ -6,11 +6,10 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface LikeRepository extends JpaRepository<Like, Long> {
+public interface LikeRepository extends JpaRepository<Like, Long> , LikeRepositoryCustom{
 
     Optional<Like> findByBoardIdAndMemberId(Long boardId, Long memberId);
 
     @Query("select count(l.id) from Like l where l.board.id=:boardId and l.member.id=:memberId")
     Long countLike(Long boardId, Long memberId);
-
 }
