+++ src/main/java/com/jjbin/mysite/api/repository/board/LikeRepositoryCustom.java
@@ -1,0 +1,7 @@
+package com.jjbin.mysite.api.repository.board;
+
+import java.util.Optional;
+
+public interface LikeRepositoryCustom {
+
+}
