+++ src/test/java/com/jjbin/mysite/api/controller/board/LikeControllerTest.java
@@ -1,0 +1,145 @@
+package com.jjbin.mysite.api.controller.board;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.jjbin.mysite.api.SessionConst;
+import com.jjbin.mysite.api.domain.Board;
+import com.jjbin.mysite.api.domain.Like;
+import com.jjbin.mysite.api.domain.Member;
+import com.jjbin.mysite.api.repository.MemberRepository;
+import com.jjbin.mysite.api.repository.board.BoardRepository;
+import com.jjbin.mysite.api.repository.board.CommentRepository;
+import com.jjbin.mysite.api.repository.board.LikeRepository;
+import com.jjbin.mysite.api.request.create.BoardCreate;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.mock.web.MockHttpSession;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+//@ExtendWith(MockitoExtension.class)
+class LikeControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Autowired
+    private ObjectMapper objectMapper;
+    @Autowired
+    private BoardRepository boardRepository;
+    @Autowired
+    private MemberRepository memberRepository;
+
+
+    @Autowired
+    private LikeRepository likeRepository;
+
+
+    @BeforeEach
+    void beforeEach(){
+        boardRepository.deleteAll();
+        memberRepository.deleteAll();
+    }
+
+
+    @Test
+    @DisplayName("게시글 좋아요 요청 - db에 좋아요 정보가 저장된다.")
+    void test1() throws Exception {
+        //given
+        MockHttpSession session = getMockHttpSession();
+        Member member = (Member) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        Board board = getTestBoard(member);
+
+        //expected
+        mockMvc.perform(post("/board/{boardId}/like", board.getId())
+                        .session(session))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.count").value(1));
+
+    }
+    @Test
+    @DisplayName("게시글 좋아요 요청 - 이미 좋아요한 경우 좋아요가 취소된다.")
+    void test1_2() throws Exception {
+        //given
+        MockHttpSession session = getMockHttpSession();
+        Member member = (Member) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        Board board = getTestBoard(member);
+
+        //when
+        likeRepository.save(
+                Like.builder()
+                        .member(member)
+                        .board(board)
+                        .build()
+        );
+        //then
+        mockMvc.perform(post("/board/{boardId}/like", board.getId())
+                        .session(session))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.count").value(0));
+
+    }
+
+    @Test
+    @DisplayName("게시글 좋아요 요청:실패 - 인증되지 않은 요청")
+    void test1_3() throws Exception {
+        //given
+        MockHttpSession session = getMockHttpSession();
+        Member member = (Member) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        Board board = getTestBoard(member);
+
+        //expected
+        mockMvc.perform(post("/board/{boardId}/like", board.getId()))
+                .andDo(print())
+                .andExpect(status().isUnauthorized())
+                .andExpect(jsonPath("$.code").value(401))
+                .andExpect(jsonPath("$.message").value("인증이 필요한 요청입니다."));
+
+
+    }
+
+    private Board getTestBoard(Member member) {
+        BoardCreate testBoardCreate = getTestBoardCreate();
+        Board testBoard = Board.createBoard(testBoardCreate, member);
+        Board board = boardRepository.save(testBoard);
+        return board;
+    }
+
+
+
+
+    private MockHttpSession getMockHttpSession() {
+
+        Member member = getTestMember();
+        MockHttpSession session = new MockHttpSession();
+        session.setAttribute(SessionConst.LOGIN_MEMBER, member);
+        return session;
+    }
+
+    private static BoardCreate getTestBoardCreate() {
+        return BoardCreate.builder()
+                .title("제목")
+                .content("내용")
+                .build();
+    }
+
+    private Member getTestMember() {
+        return memberRepository.save(
+                Member.builder()
+                        .name("이름")
+                        .loginId("아이디")
+                        .build()
+        );
+    }
+}
