+++ src/test/java/com/jjbin/mysite/api/docs/LikeDocTest.java
@@ -1,0 +1,125 @@
+package com.jjbin.mysite.api.docs;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.jjbin.mysite.api.SessionConst;
+import com.jjbin.mysite.api.domain.Board;
+import com.jjbin.mysite.api.domain.Comment;
+import com.jjbin.mysite.api.domain.Member;
+import com.jjbin.mysite.api.repository.MemberRepository;
+import com.jjbin.mysite.api.repository.board.BoardRepository;
+import com.jjbin.mysite.api.repository.board.CommentRepository;
+import com.jjbin.mysite.api.request.create.BoardCreate;
+import com.jjbin.mysite.api.request.create.CommentCreate;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.mock.web.MockHttpSession;
+import org.springframework.restdocs.RestDocumentationExtension;
+import org.springframework.test.web.servlet.MockMvc;
+
+import java.util.Optional;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
+import static org.springframework.restdocs.payload.PayloadDocumentation.*;
+import static org.springframework.restdocs.request.RequestDocumentation.*;
+import static org.springframework.restdocs.snippet.Attributes.key;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
+@AutoConfigureMockMvc
+@AutoConfigureRestDocs
+@ExtendWith(RestDocumentationExtension.class)
+class LikeDocTest {
+
+    @Autowired
+    MemberRepository memberRepository;
+    @Autowired
+    BoardRepository boardRepository;
+    @Autowired
+    CommentRepository commentRepository;
+    @Autowired
+    ObjectMapper objectMapper;
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @BeforeEach
+    void beforeEach() {
+        memberRepository.deleteAll();
+        boardRepository.deleteAll();
+        commentRepository.deleteAll();
+    }
+
+
+    @Test
+    @DisplayName("게시글 좋아요 요청")
+    void test1() throws Exception {
+        //given
+        MockHttpSession session = getMockHttpSession();
+        Member member = (Member) session.getAttribute(SessionConst.LOGIN_MEMBER);
+        Board board = getTestBoard(member);
+
+        //expected
+        mockMvc.perform(post("/board/{boardId}/like", board.getId())
+                        .session(session))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.count").value(1))
+                .andDo(document("like-board",
+                        pathParameters(
+                                parameterWithName("boardId").description("게시글 ID")
+                        ),
+                        responseFields(
+                                fieldWithPath("count").description("게시글 좋아요 갯수")
+                        )
+                ));
+    }
+
+
+
+    private Board getTestBoard(Member member) {
+        BoardCreate testBoardCreate = getTestBoardCreate();
+        Board testBoard = Board.createBoard(testBoardCreate, member);
+        Board board = boardRepository.save(testBoard);
+        return board;
+    }
+
+
+
+
+    private MockHttpSession getMockHttpSession() {
+
+        Member member = getTestMember();
+        MockHttpSession session = new MockHttpSession();
+        session.setAttribute(SessionConst.LOGIN_MEMBER, member);
+        return session;
+    }
+
+    private static BoardCreate getTestBoardCreate() {
+        return BoardCreate.builder()
+                .title("제목")
+                .content("내용")
+                .build();
+    }
+
+    private Member getTestMember() {
+        return memberRepository.save(
+                Member.builder()
+                        .name("이름")
+                        .loginId("아이디")
+                        .build()
+        );
+    }
+}
