+++ src/test/java/com/jjbin/mysite/api/service/BoardServiceTest.java
@@ -3,13 +3,12 @@ package com.jjbin.mysite.api.service;
 import com.jjbin.mysite.api.domain.Board;
 import com.jjbin.mysite.api.domain.Member;
 import com.jjbin.mysite.api.exception.ObjectNotFound;
-import com.jjbin.mysite.api.exception.Unauthorized;
 import com.jjbin.mysite.api.repository.MemberRepository;
 import com.jjbin.mysite.api.repository.board.BoardRepository;
-import com.jjbin.mysite.api.repository.mail.MailRepository;
 import com.jjbin.mysite.api.request.BoardEdit;
 import com.jjbin.mysite.api.request.SearchOption;
 import com.jjbin.mysite.api.request.create.BoardCreate;
+import com.jjbin.mysite.api.service.board.BoardService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
